Fixes for Xcode 6.1 beta.

+++ Examples/Playground/MyPlayground.playground/section-1.swift
@@ -6,7 +6,7 @@ import SKTUtils
 let pt1 = CGPoint(x: 10, y: 20)
 let pt2 = CGPoint(x: -5, y: 0)
 let pt3 = pt1 + pt2
-let pt4 = pt3 * 100
+let pt4 = pt3 * CGFloat(100)
 
 println("Point has length \(pt4.length())")
